+++ chariot-client/src/ui_state.rs
@@ -4,7 +4,6 @@ use glam::Vec2;
 use ordinal::Ordinal;
 
 use crate::{
-    application::Application,
     drawable::{
         string::StringDrawable,
         technique::{self, UILayerTechnique},
@@ -199,45 +198,5 @@ impl GraphicsManager {
             announcement_state: AnnouncementState::None,
             minimap_ui,
         }
-    }
-}
-
-impl Application {
-    pub fn render(&mut self) {
-        if let UIState::InGameHUD {
-            announcement_state, ..
-        } = &self.graphics.ui
-        {
-            match announcement_state {
-                AnnouncementState::VotingInProgress { vote_end_time, .. } => {
-                    self.graphics.make_announcement(
-                        "The audience is deciding your fate",
-                        format!(
-                            "They decide in {} seconds",
-                            (*vote_end_time - Instant::now()).as_secs()
-                        )
-                        .as_str(),
-                    );
-                }
-                AnnouncementState::VoteActiveTime {
-                    prompt: _,
-                    decision,
-                    effect_end_time,
-                } => {
-                    let effect_end_time = effect_end_time;
-                    self.graphics.make_announcement(
-                        format!("{} was chosen!", decision).as_str(),
-                        format!(
-                            "Effects will last for another {} seconds",
-                            (*effect_end_time - Instant::now()).as_secs()
-                        )
-                        .as_str(),
-                    );
-                }
-                AnnouncementState::None => {}
-            }
-        }
-
-        self.graphics.render();
     }
 }
